refactor: 승훈님 리뷰사항 적용
- 예외처리 문장 상수화
- 유틸클래스 생성 방지
- 간소화 할 수 있는 코드들 정리
