+++ src/main/java/nextstep/ladder/domain/Names.java
@@ -2,16 +2,17 @@ package nextstep.ladder.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import nextstep.ladder.exception.NotExistNameException;
 
 public class Names {
 
+    private static final String MINIMUM_PARTICIPANTS = "참가자는 최소 두명 이상부터 가능함.";
+    private static final String NO_PARTICIPANTS = "찾으려는 참가자 이름이 없음.";
     private List<Name> names;
 
     public Names(List<String> names) {
         if (names.size() < 2) {
-            throw new IllegalArgumentException("참가자는 최소 두명 이상부터 가능함.");
+            throw new IllegalArgumentException(MINIMUM_PARTICIPANTS);
         }
 
         this.names = names.stream()
@@ -25,17 +26,21 @@ public class Names {
             .collect(Collectors.toUnmodifiableList());
     }
 
-    public int indexOf(final String name) {
-        return IntStream.range(0, names.size())
-            .filter(value -> names.get(value).get().equals(name))
-            .boxed()
-            .findAny()
-            .orElseThrow(() -> new NotExistNameException("찾으려는 이름이 없음"));
+    public int indexOf(final Name name) {
+        if (!names.contains(name)) {
+            throw new NotExistNameException(NO_PARTICIPANTS);
+        }
+
+        return names.indexOf(name);
     }
 
     @Override
     public String toString() {
         return "Names{" + names +
             "}\n";
+    }
+
+    public int size() {
+        return names.size();
     }
 }
