+++ src/test/java/nextstep/ladder/domain/NamesTest.java
@@ -25,16 +25,16 @@ class NamesTest {
     @DisplayName("목록에 있는 이름의 인덱스를 찾는다.")
     @Test
     void indexOfTest1() {
-        int index = this.names.indexOf(name1);
+        int index = this.names.indexOf(new Name(name1));
         assertThat(index).isEqualTo(0);
-        index = this.names.indexOf(name2);
+        index = this.names.indexOf(new Name(name2));
         assertThat(index).isEqualTo(1);
     }
 
     @DisplayName("목록에 없는 이름의 인덱스를 찾으면 예외가 발생한다.")
     @Test
     void indexOfTest2() {
-        assertThatThrownBy(() -> this.names.indexOf("1234"))
+        assertThatThrownBy(() -> this.names.indexOf(new Name("1234")))
             .isInstanceOf(NotExistNameException.class);
     }
